Update chart on TDD table hover
