+++ web-common/src/features/dashboards/time-dimension-details/charts/utils.ts
@@ -3,15 +3,8 @@ import {
   buildVegaLiteSpec,
 } from "@rilldata/web-common/features/charts/templates/build-template";
 import { TDDChartMap } from "@rilldata/web-common/features/charts/types";
-import { COMPARIONS_COLORS } from "@rilldata/web-common/features/dashboards/config";
-import {
-  MainAreaColorGradientDark,
-  MainLineColor,
-} from "@rilldata/web-common/features/dashboards/time-series/chart-colors";
 import type { DimensionDataItem } from "@rilldata/web-common/features/dashboards/time-series/multiple-dimension-queries";
-import { TIME_GRAIN } from "@rilldata/web-common/lib/time/config";
-import { V1TimeGrain } from "@rilldata/web-common/runtime-client";
-import { VisualizationSpec } from "svelte-vega";
+import { View, VisualizationSpec } from "svelte-vega";
 import { TDDAlternateCharts, TDDChart } from "../types";
 
 export function reduceDimensionData(dimensionData: DimensionDataItem[]) {
@@ -76,144 +69,63 @@ export function getVegaSpecForTDD(
   return spec;
 }
 
-function patchSpecForTimeComparison(
-  sanitizedSpec,
-  chartType: TDDAlternateCharts,
-  timeUnit: string,
-  measureName: string,
-  yEncoding,
-  colorEncoding,
-) {
-  yEncoding.field = "measure";
+function isSignalEqual(currentValues, newValues) {
+  if (currentValues.length !== newValues.length) {
+    return false;
+  }
 
-  sanitizedSpec.transform = [
-    // Sanitize and transform comparison data in the right time format
-    { timeUnit: timeUnit, field: "comparison\\.ts", as: "comparison_ts" },
-    // Expand datum to have a key field to differentiate between current and comparison data
-    { fold: ["ts", "comparison_ts"] },
-    // Add a measure field to hold the right measure value
-    {
-      calculate: `(datum['key'] === 'comparison_ts' ? datum['comparison.${measureName}'] : datum['${measureName}'])`,
-      as: "measure",
-    },
-    // Add a time field to hold the right time value
-    {
-      calculate:
-        "(datum['key'] === 'comparison_ts' ? datum['comparison_ts'] : datum['ts'])",
-      as: "time",
-    },
-  ];
-
-  colorEncoding.scale = {
-    domain: ["ts", "comparison_ts"],
-    range: [MainLineColor, MainAreaColorGradientDark],
-  };
-
-  if (chartType === TDDChart.STACKED_AREA) {
-    /**
-     * For stacked area charts, we don't need to pivot transform as the
-     * comparison values are already in the right format.
-     */
-
-    // TODO: This is error prone, find a better way to do this without
-    // mutating the template
-    const stackedAreaPivotLayer = sanitizedSpec.layer[2];
-
-    if (stackedAreaPivotLayer) {
-      delete stackedAreaPivotLayer.transform;
+  for (let i = 0; i < currentValues.length; i++) {
+    if (currentValues[i] !== newValues[i]) {
+      return false;
     }
   }
+  return true;
 }
 
-export function patchSpecForTDD(
-  spec,
+function hasChartDimensionParam(
   chartType: TDDAlternateCharts,
-  timeGrain: V1TimeGrain,
-  xMin: Date,
-  xMax: Date,
-  isTimeComparison: boolean,
-  measureName: string,
-  selectedDimensionValues: (string | null)[] = [],
-): VisualizationSpec {
-  if (!spec) return spec;
-
-  /**
-   * Sub level types are not being exported from the vega-lite package.
-   * This makes it hard to modify the specs without breaking typescript
-   * interface. For now we have removed the types for the spec and will
-   * add them back when we have the support for it.
-   * More at https://github.com/vega/vega-lite/issues/9222
-   */
-
-  const sanitizedSpec = structuredClone(spec);
-  let xEncoding;
-  let yEncoding;
-  let colorEncoding;
-  if (sanitizedSpec.encoding) {
-    xEncoding = sanitizedSpec.encoding.x;
-    yEncoding = sanitizedSpec.encoding.y;
-
-    colorEncoding = sanitizedSpec.encoding?.color;
-
-    xEncoding.scale = {
-      domain: [xMin.toISOString(), xMax.toISOString()],
-    };
+  hasComparison: boolean,
+) {
+  if (!hasComparison) {
+    return false;
+  } else if (chartType === TDDChart.STACKED_AREA) {
+    return false;
   }
 
-  if (!xEncoding || !yEncoding) {
-    return sanitizedSpec;
-  }
-
-  const selectedValuesLength = selectedDimensionValues.length;
-  if (colorEncoding && selectedValuesLength) {
-    colorEncoding.scale = {
-      domain: selectedDimensionValues,
-      range: COMPARIONS_COLORS.slice(0, selectedValuesLength),
-    };
-  }
-
-  // Set extents for x-axis
-  xEncoding.scale = {
-    domain: [xMin.toISOString(), xMax.toISOString()],
-  };
-
-  const timeLabelFormat = TIME_GRAIN[timeGrain]?.d3format as string;
-  // Remove titles from axes
-  xEncoding.axis = {
-    ticks: false,
-    orient: "top",
-    title: "",
-    formatType: "time",
-    format: timeLabelFormat,
-  };
-  yEncoding.axis = { title: "", formatType: "measureFormatter" };
-
-  // Set timeUnit for x-axis using timeGrain
-  const timeUnit = timeGrainToVegaTimeUnitMap[timeGrain];
-  xEncoding.timeUnit = timeUnit;
-
-  if (isTimeComparison) {
-    patchSpecForTimeComparison(
-      sanitizedSpec,
-      chartType,
-      timeUnit,
-      measureName,
-      yEncoding,
-      colorEncoding,
-    );
-  }
-
-  return sanitizedSpec;
+  return true;
 }
 
-const timeGrainToVegaTimeUnitMap = {
-  [V1TimeGrain.TIME_GRAIN_SECOND]: "yearmonthdatehoursminutesseconds",
-  [V1TimeGrain.TIME_GRAIN_MINUTE]: "yearmonthdatehoursminutes",
-  [V1TimeGrain.TIME_GRAIN_HOUR]: "yearmonthdatehours",
-  [V1TimeGrain.TIME_GRAIN_DAY]: "yearmonthdate",
-  [V1TimeGrain.TIME_GRAIN_WEEK]: "yearweek",
-  [V1TimeGrain.TIME_GRAIN_MONTH]: "yearmonth",
-  [V1TimeGrain.TIME_GRAIN_QUARTER]: "yearquarter",
-  [V1TimeGrain.TIME_GRAIN_YEAR]: "year",
-  [V1TimeGrain.TIME_GRAIN_UNSPECIFIED]: "yearmonthdate",
-};
+export function updateVegaOnTableHover(
+  viewVL: View | undefined,
+  chartType: TDDAlternateCharts,
+  hasComparison: boolean,
+  time: Date | null | undefined,
+  dimensionValue: string | null | undefined,
+) {
+  if (!viewVL) {
+    return;
+  }
+
+  const epochTime = time ? time.getTime() : null;
+  const hasDimensionParam = hasChartDimensionParam(chartType, hasComparison);
+  const values = epochTime
+    ? hasDimensionParam
+      ? [epochTime, dimensionValue]
+      : [epochTime]
+    : null;
+
+  const newValue = epochTime
+    ? {
+        unit: "",
+        fields: viewVL.signal("hover_tuple_fields"),
+        values,
+      }
+    : null;
+  const currentValues = (viewVL.signal("hover_tuple") || { values: [] }).values;
+  const newValues = values || [];
+  if (isSignalEqual(currentValues, newValues)) {
+    return;
+  }
+  viewVL.signal("hover_tuple", newValue);
+  viewVL.runAsync();
+}
